+++ turn/src/executor/bridge_ledger.rs
@@ -344,7 +344,7 @@ impl TurnExecutor {
                 .state
                 .get_field(LIVE_FIELD)
                 .expect("fixed slot 1 always present");
-            journal.record_set_field(req.ledger_cell, LIVE_FIELD, Some(old_live));
+            journal.record_set_field(req.ledger_cell, LIVE_FIELD as u64, Some(old_live));
             cell.state.set_field(LIVE_FIELD, encode_u64(new_live));
         }
 
@@ -449,7 +449,7 @@ impl TurnExecutor {
                 .state
                 .get_field(LOCKED_FIELD)
                 .expect("fixed slot 0 always present");
-            journal.record_set_field(req.ledger_cell, LOCKED_FIELD, Some(old_locked));
+            journal.record_set_field(req.ledger_cell, LOCKED_FIELD as u64, Some(old_locked));
             cell.state.set_field(LOCKED_FIELD, encode_u64(new_locked));
         }
